+++ Modules/Intent.Modules.EntityFrameworkCore.Repositories/Templates/Repository/RepositoryTemplatePartial.cs
@@ -35,6 +35,8 @@ namespace Intent.Modules.EntityFrameworkCore.Repositories.Templates.Repository
         public RepositoryTemplate(IOutputTarget outputTarget, ClassModel model)
             : base(TemplateId, outputTarget, model)
         {
+            FulfillsRole(TemplateRoles.Repository.Implementation.Entity);
+
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .AddUsing("System")
                 .AddUsing("System.Linq")