+++ Modules/Intent.Modules.MongoDb.Repositories/Templates/Repository/RepositoryTemplatePartial.cs
@@ -8,6 +8,7 @@ using Intent.Modules.Common.CSharp.Builder;
 using Intent.Modules.Common.CSharp.DependencyInjection;
 using Intent.Modules.Common.CSharp.Templates;
 using Intent.Modules.Common.Templates;
+using Intent.Modules.Constants;
 using Intent.Modules.Entities.Repositories.Api.Templates;
 using Intent.Modules.Entities.Repositories.Api.Templates.EntityRepositoryInterface;
 using Intent.Modules.Modelers.Domain.Settings;
@@ -28,6 +29,8 @@ namespace Intent.Modules.MongoDb.Repositories.Templates.Repository
         [IntentManaged(Mode.Fully, Body = Mode.Ignore)]
         public RepositoryTemplate(IOutputTarget outputTarget, ClassModel model) : base(TemplateId, outputTarget, model)
         {
+            FulfillsRole(TemplateRoles.Repository.Implementation.Entity);
+
             CSharpFile = new CSharpFile(this.GetNamespace(), this.GetFolderPath())
                 .AddUsing("System")
                 .AddUsing("System.Collections.Generic")