+++ Modules/Intent.Modules.Dapr.AspNetCore.StateManagement/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.2.10
+
+- Improvement: Added support for extending an entity repository as well as generating custom entity-less repositories.
+
 ### Version 1.2.9
 
 - Improvement: Inline with the update to `Intent.Entities.Repositories.Api` version `5.1.4`, removed 2nd type parameter for repository interface.